refactor(contracts): split registry surface by concern
